+++ packages/forma-36-react-components/src/components/DateTime/index.ts
@@ -1,8 +1,4 @@
 import { DateTime } from './DateTime/DateTime';
 import { RelativeDate } from './RelativeDate/RelativeDate';
 import * as datetime from './dateUtils';
-export {
-  datetime,
-  DateTime,
-  RelativeDate
-}
+export { datetime, DateTime, RelativeDate };