+++ packages/forma-36-react-components/src/components/Icon/svg/ChevronDown.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+function SvgChevronDown(props: React.SVGProps<SVGSVGElement>) {
+  return (
+    <svg height="1em" viewBox="0 0 24 24" width="1em" {...props}>
+      <path d="M7.41 7.84L12 12.42l4.59-4.58L18 9.25l-6 6-6-6z" />
+      <path d="M0-.75h24v24H0z" fill="none" />
+    </svg>
+  );
+}
+
+export default SvgChevronDown;