+++ packages/forma-36-react-components/src/components/Icon/svg/References.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+function SvgReferences(props: React.SVGProps<SVGSVGElement>) {
+  return (
+    <svg height="1em" viewBox="0 0 24 24" width="1em" {...props}>
+      <path fill="none" d="M0 0h24v24H0z" />
+      <path d="M22 11V3h-7v3H9V3H2v8h7V8h2v10h4v3h7v-8h-7v3h-2V8h2v3h7zM7 9H4V5h3v4zm10 6h3v4h-3v-4zm0-10h3v4h-3V5z" />
+    </svg>
+  );
+}
+
+export default SvgReferences;